Address presence review findings
Move protocol-visible presence state behind adapter-owned connection
identifiers and validate presence payloads before mutating the shared
registry. This keeps transport IDs out of `Session` while preserving
deterministic fan-out and disconnect notifications.

Tighten field-300 duplicate validation, fix Notify Change User flag
encoding, update transaction policy coverage, and align the presence
docs and roadmap with the delivered 300-304 scope.
