+++ docs/users-guide.md
@@ -58,14 +58,15 @@ in `mxd::server::cli`, while the active networking runtime is selected by the
 - Compatibility behaviour is unchanged by roadmap item 1.5.4; that milestone
   adds bounded Kani verification for XOR round-trips and login
   sub-version/version gating invariants.
-- Presence parity is now available in the wireframe server. Successful login
-  immediately makes a default user visible to peers, `Get User Name List`
-  (300) returns the current online roster, `Set Client User Info` (304)
-  updates nickname/icon/options for the current session and informs peers via
-  `Notify Change User` (301), and disconnects remove the user via `Notify
-  Delete User` (302). Until richer account metadata exists, the initial
-  visible nickname is the account username, icon `0`, and `Get Client Info
-  Text` (303) returns a blank info string alongside the current visible name.
+- Presence parity is now available in the wireframe server. Accounts that can
+  skip agreement become visible to peers immediately after login. Agreement-
+  gated accounts enter `PendingAgreement` and become visible only after the
+  session reaches `Online`, such as after Agreement Acceptance (121) finalizes.
+  `Get User Name List` (300) and `Notify Change User` (301) are emitted only
+  once the session is `Online`. Once online, `Notify Delete User` (302),
+  `Get Client Info Text` (303), and `Set Client User Info` (304) behave as
+  before: disconnects remove the user, info lookup returns the visible name
+  with blank info text, and session nickname/icon/options updates notify peers.
 - Internal release validation uses
   `docs/internal-compatibility-matrix.md` as the compatibility source of truth.
   Release-note QA sign-off must reference that matrix using the checklist in
