+++ src/handler.rs
@@ -2,7 +2,15 @@
 //!
 //! The handler owns per-client [`Session`] state and dispatches incoming
 //! transactions to [`Command`] processors.
-use std::{error::Error, fmt, net::SocketAddr, sync::Arc};
+use std::{
+    error::Error,
+    fmt,
+    net::SocketAddr,
+    sync::{
+        Arc,
+        atomic::{AtomicU64, Ordering},
+    },
+};
 
 use argon2::Argon2;
 
@@ -16,6 +24,8 @@ use crate::{
     transaction::{Transaction, parse_transaction},
 };
 
+static NEXT_LEGACY_PRESENCE_CONNECTION_ID: AtomicU64 = AtomicU64::new(1);
+
 /// Per-connection context used by `handle_request`.
 #[derive(Clone)]
 pub struct Context {
@@ -27,6 +37,8 @@ pub struct Context {
     pub argon2: Arc<Argon2<'static>>,
     /// Shared presence registry for legacy request processing.
     pub presence: Arc<PresenceRegistry>,
+    /// Adapter-owned identifier used when publishing this connection's presence.
+    pub presence_connection_id: OutboundConnectionId,
 }
 
 /// Session state for a single connection.
@@ -54,8 +66,6 @@ pub struct Session {
     pub connection_flags: ConnectionFlags,
     /// Automatic response text associated with the current session.
     pub auto_response: Option<String>,
-    /// Outbound identifier for presence notifications.
-    pub outbound_connection_id: Option<OutboundConnectionId>,
 }
 
 /// Error returned when a privilege check fails.
@@ -171,9 +181,11 @@ impl Session {
 
     /// Build a public presence snapshot when the session is online and visible.
     #[must_use]
-    pub fn presence_snapshot(&self) -> Option<PresenceSnapshot> {
+    pub fn presence_snapshot(
+        &self,
+        connection_id: OutboundConnectionId,
+    ) -> Option<PresenceSnapshot> {
         let user_id = self.user_id?;
-        let connection_id = self.outbound_connection_id?;
         if !self.shows_in_user_list() {
             return None;
         }
@@ -207,12 +219,13 @@ impl Context {
             pool,
             argon2,
             presence: Arc::new(PresenceRegistry::default()),
+            presence_connection_id: next_legacy_presence_connection_id(),
         }
     }
 
     /// Create a new connection context with shared presence.
     #[must_use]
-    pub const fn with_presence(
+    pub fn with_presence(
         peer: SocketAddr,
         pool: DbPool,
         argon2: Arc<Argon2<'static>>,
@@ -223,8 +236,14 @@ impl Context {
             pool,
             argon2,
             presence,
+            presence_connection_id: next_legacy_presence_connection_id(),
         }
     }
+}
+
+fn next_legacy_presence_connection_id() -> OutboundConnectionId {
+    let id = NEXT_LEGACY_PRESENCE_CONNECTION_ID.fetch_add(1, Ordering::Relaxed);
+    OutboundConnectionId::new(id)
 }
 
 /// Parse and handle a single request frame without performing network I/O.
@@ -244,157 +263,11 @@ pub async fn handle_request(
         pool: ctx.pool.clone(),
         session,
         presence: ctx.presence.as_ref(),
+        presence_connection_id: Some(ctx.presence_connection_id),
     })
     .await
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::wireframe::test_helpers::dummy_pool;
-
-    #[tokio::test]
-    async fn context_carries_shared_argon2_reference() {
-        let pool = dummy_pool();
-        let argon2 = Arc::new(Argon2::default());
-        let peer: SocketAddr = "127.0.0.1:9001".parse().expect("loopback address");
-
-        let ctx = Context::new(peer, pool, Arc::clone(&argon2));
-
-        assert!(Arc::ptr_eq(&ctx.argon2, &argon2));
-        assert_eq!(Arc::strong_count(&argon2), 2);
-        assert_eq!(ctx.peer, peer);
-    }
-
-    #[tokio::test]
-    async fn multiple_contexts_share_single_argon2_instance() {
-        let pool = dummy_pool();
-        let argon2 = Arc::new(Argon2::default());
-
-        let ctx_a = Context::new(
-            "127.0.0.1:9002".parse().expect("loopback"),
-            pool.clone(),
-            Arc::clone(&argon2),
-        );
-        let ctx_b = Context::new(
-            "127.0.0.1:9003".parse().expect("loopback"),
-            pool,
-            Arc::clone(&argon2),
-        );
-
-        assert!(Arc::ptr_eq(&ctx_a.argon2, &argon2));
-        assert!(Arc::ptr_eq(&ctx_b.argon2, &argon2));
-        assert_eq!(Arc::strong_count(&argon2), 3);
-
-        drop(ctx_a);
-        assert_eq!(Arc::strong_count(&argon2), 2);
-        drop(ctx_b);
-        assert_eq!(Arc::strong_count(&argon2), 1);
-    }
-
-    #[test]
-    fn session_default_is_unauthenticated() {
-        let session = Session::default();
-        assert!(!session.is_authenticated());
-        assert!(session.privileges.is_empty());
-        assert_eq!(session.phase, SessionPhase::Unauthenticated);
-        assert!(session.display_name.is_empty());
-        assert_eq!(session.icon_id, 0);
-        assert!(session.connection_flags.is_empty());
-        assert!(session.auto_response.is_none());
-        assert!(session.outbound_connection_id.is_none());
-    }
-
-    #[test]
-    fn session_is_authenticated_with_user_id() {
-        let session = Session {
-            user_id: Some(42),
-            ..Default::default()
-        };
-        assert!(session.is_authenticated());
-        assert!(!session.is_online());
-    }
-
-    #[test]
-    fn session_has_privilege_returns_true_when_present() {
-        let session = Session {
-            user_id: Some(1),
-            privileges: Privileges::DOWNLOAD_FILE,
-            ..Default::default()
-        };
-        assert!(session.has_privilege(Privileges::DOWNLOAD_FILE));
-    }
-
-    #[test]
-    fn session_has_privilege_returns_false_when_absent() {
-        let session = Session {
-            user_id: Some(1),
-            privileges: Privileges::DOWNLOAD_FILE,
-            ..Default::default()
-        };
-        assert!(!session.has_privilege(Privileges::UPLOAD_FILE));
-    }
-
-    #[test]
-    fn session_require_privilege_fails_when_unauthenticated() {
-        let session = Session::default();
-        let result = session.require_privilege(Privileges::DOWNLOAD_FILE);
-        assert_eq!(result, Err(PrivilegeError::NotAuthenticated));
-    }
-
-    #[test]
-    fn session_require_privilege_fails_when_missing_privilege() {
-        let session = Session {
-            user_id: Some(1),
-            privileges: Privileges::READ_CHAT,
-            ..Default::default()
-        };
-        let result = session.require_privilege(Privileges::DOWNLOAD_FILE);
-        assert_eq!(
-            result,
-            Err(PrivilegeError::InsufficientPrivileges(
-                Privileges::DOWNLOAD_FILE
-            ))
-        );
-    }
-
-    #[test]
-    fn session_require_privilege_succeeds_when_present() {
-        let session = Session {
-            user_id: Some(1),
-            privileges: Privileges::DOWNLOAD_FILE | Privileges::READ_CHAT,
-            ..Default::default()
-        };
-        let result = session.require_privilege(Privileges::DOWNLOAD_FILE);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn session_require_authenticated_fails_when_unauthenticated() {
-        let session = Session::default();
-        let result = session.require_authenticated();
-        assert_eq!(result, Err(PrivilegeError::NotAuthenticated));
-    }
-
-    #[test]
-    fn session_require_authenticated_succeeds_when_logged_in() {
-        let session = Session {
-            user_id: Some(1),
-            ..Default::default()
-        };
-        let result = session.require_authenticated();
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn privilege_error_display_not_authenticated() {
-        let err = PrivilegeError::NotAuthenticated;
-        assert_eq!(err.to_string(), "authentication required");
-    }
-
-    #[test]
-    fn privilege_error_display_insufficient_privileges() {
-        let err = PrivilegeError::InsufficientPrivileges(Privileges::DOWNLOAD_FILE);
-        assert!(err.to_string().contains("insufficient privileges"));
-    }
-}
+#[path = "handler_tests.rs"]
+mod tests;
