+++ src/wireframe/routes/mod.rs
@@ -38,7 +38,7 @@ use crate::wireframe::codec::HotlineTransaction;
 use crate::{
     db::DbPool,
     presence::PresenceRegistry,
-    server::outbound::OutboundMessaging,
+    server::outbound::{OutboundConnectionId, OutboundMessaging},
     transaction::FrameHeader,
     wireframe::router::{RouteContext as RouterRouteContext, WireframeRouter},
 };
@@ -166,6 +166,7 @@ pub(crate) struct TransactionMiddleware {
     peer: SocketAddr,
     messaging: Arc<dyn OutboundMessaging>,
     presence: Arc<PresenceRegistry>,
+    presence_connection_id: OutboundConnectionId,
 }
 
 /// Construction parameters for [`TransactionMiddleware`].
@@ -182,6 +183,8 @@ pub(crate) struct TransactionMiddlewareConfig {
     pub(crate) messaging: Arc<dyn OutboundMessaging>,
     /// Shared online presence registry.
     pub(crate) presence: Arc<PresenceRegistry>,
+    /// Adapter-owned connection identifier for presence snapshots.
+    pub(crate) presence_connection_id: OutboundConnectionId,
 }
 
 impl TransactionMiddleware {
@@ -195,6 +198,7 @@ impl TransactionMiddleware {
             peer: config.peer,
             messaging: config.messaging,
             presence: config.presence,
+            presence_connection_id: config.presence_connection_id,
         }
     }
 }
@@ -208,6 +212,7 @@ struct TransactionHandler {
     peer: SocketAddr,
     messaging: Arc<dyn OutboundMessaging>,
     presence: Arc<PresenceRegistry>,
+    presence_connection_id: OutboundConnectionId,
 }
 
 #[async_trait]
@@ -226,6 +231,7 @@ impl Service for TransactionHandler {
                         session: &mut session_guard,
                         messaging: self.messaging.as_ref(),
                         presence: self.presence.as_ref(),
+                        presence_connection_id: self.presence_connection_id,
                     },
                 )
                 .await
@@ -253,6 +259,7 @@ impl Transform<HandlerService<Envelope>> for TransactionMiddleware {
             peer: self.peer,
             messaging: Arc::clone(&self.messaging),
             presence: Arc::clone(&self.presence),
+            presence_connection_id: self.presence_connection_id,
         };
         HandlerService::from_service(id, wrapped)
     }
