+++ src/wireframe/routes/tests/mod.rs
@@ -3,4 +3,5 @@
 mod error_cases;
 mod helpers;
 mod middleware_cases;
+mod presence_routing_cases;
 mod routing_cases;
